Basic rooms implemented
Rooms can be linked together easily with a script, and there is a base
room that does all the necessary things to set a room up that all
other rooms should inherit. The parser object can understand basic
direction inputs for moving between rooms and actually move to
another room, with description text being printed out. I'm sure there
is going to be more to do, but this is a nice first step.